Move kilterboard component to its own directory
