+++ my-react-app/src/KilterBoard/KilterBoardLoader.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect, useMemo } from "react";
 import { useParams } from "react-router-dom";
 import KilterBoard from "./KilterBoard";
-import { fetchBoardDetails } from "./api";
-import { boardLayouts } from "./board-data";
+import { fetchBoardDetails } from "../api";
+import { boardLayouts } from "../board-data";
 
 // TODO: Unhardcode set_ids
 const KilterBoardLoader = ({ litUpHolds }) => {
